// backend/src/common/excel/types/supplier-excel-row.types.ts
export interface SupplierExcelRow {
  supplierCode: string;

  name: string;

  mobile: string;

  alternateMobile?: string;

  email?: string;

  gstNo?: string;

  address?: string;

  city?: string;

  state?: string;

  pinCode?: string;

  openingBalance: number;

  notes?: string;

  isActive: boolean;
}